requirements.txt
aws-cdk-lib==2.23.0
boto3==1.18.53
cdk-nag==2.12.46
constructs>=10.0.0,<11.0.0
